bugs fixed, welcome msg added handles stored as lowercase

+++ discordbot.py
@@ -1,4 +1,5 @@
 import discord
+import re
 from discord import app_commands
 from discord.ext import commands
 from globals_ import *
@@ -10,6 +11,14 @@ intents.message_content = True
 intents.typing = False
 intents.presences = False
 
+def create_start_message():
+    embed = discord.Embed(title='Hello World!', description='The bot is now online and ready to serve you.', color=0x0000ff)
+    embed.set_thumbnail(url='https://i.imgur.com/GyWdKAx.jpg')
+    embed.add_field(name='Commands', value='!help - Shows a list of available commands', inline=False)
+    return embed
+
+def to_lower(string):
+    return string.lower()
     
 async def load_database(stream,channel):
     # Execute a COUNT statement to count the number of rows in the table
@@ -27,19 +36,22 @@ async def load_database(stream,channel):
 class Tracker(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
-        self.stream = None
         
     @commands.hybrid_command()
-    async def add_user(self,ctx, handle: str, question: str):
+    async def add_user(self,ctx, handle: to_lower, question: str):
     
+        if not re.match(TWITTER_HANDLE_REGEX, handle):
+            await ctx.send("Invalid handle")
+            return;
+        
         # Check if handle already exists in the database
         if await handle_exist(handle):
             await ctx.send('Handle already exists in database')
             return
     
         try:
-            await self.stream.add_handle(handle)
-            self.stream.custom_filter()
+            await self.bot.stream.add_handle(handle)
+            self.bot.stream.custom_filter()
             CURSOR.execute("INSERT INTO users VALUES (?, ?)", (handle, question))
             cnx.commit()
             await ctx.send(f"Tracking {handle} for question: {question}")
@@ -48,15 +60,19 @@ class Tracker(commands.Cog):
     
         
     @commands.hybrid_command()
-    async def remove_user(self,ctx, handle: str):
-
+    async def remove_user(self,ctx, handle: to_lower):
+        
+        if not re.match(TWITTER_HANDLE_REGEX, handle):
+            await ctx.send("Invalid handle")
+            return
+        
         # Check if handle already exists in the database
         if not (await handle_exist(handle)):
             await ctx.send(f"User is not currently tracked")
             return
     
-        await self.stream.remove_handle(handle)
-        self.stream.custom_filter()
+        await self.bot.stream.remove_handle(handle)
+        self.bot.stream.custom_filter()
         CURSOR.execute("DELETE FROM users WHERE handle = ?", (handle,))
         cnx.commit()
         await ctx.send(f"Stopped tracking {handle}")
@@ -80,15 +96,15 @@ class Tracker(commands.Cog):
     async def start(self,ctx):
         # Start the bot
         await ctx.send('Starting bot')
-        self.stream = MyStreamListener(self.channel,CURSOR)
-        await load_database(self.stream,self.channel)
+        self.bot.stream = MyStreamListener(self.bot.channel,CURSOR)
+        await load_database(self.bot.stream,self.bot.channel)
     
     
     @commands.hybrid_command()
     async def stop(self,ctx):
         # Stop the bot
         await ctx.send('Stopping bot')
-        self.stream.disconnect()
+        self.bot.stream.disconnect()
 
 
 class DiscordBot(commands.Bot):
@@ -101,8 +117,6 @@ class DiscordBot(commands.Bot):
         
     async def setup_hook(self):
         tracker = Tracker(self)
-        tracker.stream = self.stream
-        tracker.channel = self.channel
         await self.add_cog(tracker)
         
         
@@ -119,5 +133,7 @@ class DiscordBot(commands.Bot):
         if self.stream is None:
             self.stream = MyStreamListener(self.channel,CURSOR)
             await load_database(self.stream,self.channel)
+            
+        await self.channel.send(embed=create_start_message())
